+++ tests/testthat/test_dtwSat.R
@@ -21,8 +21,8 @@
 names(patterns.list)
 
 # Perform twdtw alignment
-weight.fun = function(x, y) 0.1*x
-alig = twdtw(patterns=patterns.list["Soybean"], timeseries=template, step.matrix = symmetric1, 
+weight.fun = function(phi, psi) 0.1*phi
+alig = twdtw(x=template, patterns=patterns.list["Soybean"], step.matrix = symmetric1, 
              dist.method = "Euclidean", weight.fun = weight.fun, keep=TRUE)
 
 is(alig, "twdtw")
@@ -41,7 +41,7 @@ gp2
 
 
 # Wavelet filter
-sy = waveletSmoothing(timeseries = template, frequency=8, wf = "la8", J=1, 
+sy = waveletSmoothing(x = template, frequency=8, wf = "la8", J=1, 
                       boundary = "periodic")
 
 
@@ -60,16 +60,16 @@ data.frame(Old.Length=sapply(patterns.list, nrow), New.length=sapply(new.pattern
 
 # Perform twdtw to patterns list 
 weight.fun = logisticWeight(alpha=-0.1, beta=100, theta=0.5)
-malig = twdtw(patterns=new.patterns.list, timeseries=template,
+alig = twdtw(x=template, patterns=new.patterns.list, 
               weight.fun = weight.fun, keep=TRUE)
 
 # Classify interval
-best_class = classifyIntervals(x=malig, from=as.Date("2009-09-01"), 
+best_class = classifyIntervals(x=alig, from=as.Date("2009-09-01"), 
                               to=as.Date("2013-09-01"), by = "6 month",
                               overlap=.3, threshold=Inf)
 best_class
 
-gp = plotGroup(x=malig, from=as.Date("2009-09-01"),  
+gp = plotGroup(x=alig, from=as.Date("2009-09-01"),  
               to=as.Date("2013-09-01"), by = "6 month",
               overlap=.3)
 gp
@@ -79,24 +79,37 @@ gp
 
 
 # Perform twdtw to patterns list 
-malig = twdtw(patterns = patterns.list, timeseries = template.list[[2]], 
-              weight.fun = weight.fun, normalize.patterns=TRUE, patterns.length = 23,
+alig = twdtw(x=template.list[[2]], patterns = patterns.list, 
+             weight.fun = weight.fun, 
+             normalize.patterns=TRUE, patterns.length = 23,
               keep=TRUE)
 
 # Classify interval
-best_class = classifyIntervals(x=malig, from=as.Date("2007-09-01"), 
+best_class = classifyIntervals(x=alig, from=as.Date("2007-09-01"), 
                                to=as.Date("2013-09-01"), by = "6 month",
                                overlap=.3, threshold=Inf)
 best_class
 
-gp = plotGroup(x=malig, from=as.Date("2007-09-01"),  
+gp = plotGroup(x=alig, from=as.Date("2007-09-01"),  
                to=as.Date("2013-09-01"), by = "6 month",
                overlap=.3)
 gp
 
 
+# Apply twdtw to a list of time series 
+aligs = lapply(template.list, twdtw, patterns = patterns.list, 
+             weight.fun = weight.fun, normalize.patterns=TRUE, 
+             patterns.length = 23, keep=TRUE)
+
+gp.list = lapply(aligs, plotGroup, from=as.Date("2007-09-01"),  
+               to=as.Date("2013-09-01"), by = "6 month", overlap=.3,
+               attr="evi")
+
+grid.arrange(grobs=gp.list, ncol=1)
+
+
 # Plot cost matrix 
-alig = twdtw(patterns.list[["Soybean"]], timeseries=template, weight.fun = weight.fun, 
+alig = twdtw(x=template, patterns.list[["Soybean"]], weight.fun = weight.fun, 
              keep=TRUE)
  
 
@@ -112,15 +125,16 @@ grid.arrange(gp)
  
 # Test bands order 
 patterns.list2 = lapply(patterns.list, function(qq) qq[,c("evi"), drop=FALSE])
-malig = twdtw(patterns = patterns.list2, timeseries = template.list[[2]], 
-              weight.fun = weight.fun, normalize.query=TRUE, query.length = 23, keep=TRUE)
+alig = twdtw( x = template.list[[2]], patterns = patterns.list2, 
+              weight.fun = weight.fun, normalize.patterns=TRUE,
+              patterns.length = 23, keep=TRUE)
 # Classify interval
-best_class = classifyIntervals(x=malig, from=as.Date("2007-09-01"), 
+best_class = classifyIntervals(x=alig, from=as.Date("2007-09-01"), 
                                to=as.Date("2013-09-01"), by = "6 month",
                                overlap=.3, threshold=Inf)
 best_class
 
-gp = plotGroup(x=malig, from=as.Date("2007-09-01"),  
+gp = plotGroup(x=alig, from=as.Date("2007-09-01"),  
                to=as.Date("2013-09-01"), by = "6 month",
                overlap=.3)
 gp